backend/requirements.txt
python-dotenv
requests
twitchio
twitchAPI
websockets